refactor(backend): extract auth flow module
